Extract sourcehub cosmos DAC tests to sibling module
Move the oversized inline test module from cosmos/dac.rs into
dac/tests.rs, matching the existing cosmos/tx test layout, without
changing DAC authorization, mock, or assertion behavior.
